Fixed eslint error
